gpio: Delay adding buttons to the Button map
If a platform doesn't have all the buttons defined yet (e.g. because
it's in the process of being brought up), the PressButton rpc currently
hangs and never returns. I would have expected some sort of error message
and indeed the error handler for the button accessor:
```
	c, ok := g.Button[b]
	if !ok {
		return nil, fmt.Errorf("Unknown button %v", b)
	}
```
suggests that such an error message was indeed intended. However,
at the moment we're initializing the map for all buttons, so
`ok` is true, even if no button has ever been registered.
Switch that to not set up the map until the button was actually
registered, thus allowing above error case to fire.